修复Windows下编译endless失败问题
