+++ server/core/server_other.go
@@ -1,0 +1,17 @@
+// +build !windows
+
+package core
+
+import (
+	"github.com/fvbock/endless"
+	"github.com/gin-gonic/gin"
+	"time"
+)
+
+func initServer(address string, router *gin.Engine) server {
+	s := endless.NewServer(address, router)
+	s.ReadHeaderTimeout = 10 * time.Millisecond
+	s.WriteTimeout = 10 * time.Second
+	s.MaxHeaderBytes = 1 << 20
+	return s
+}
